requirements.txt
# Web Framework and API
Flask==3.1.0
flask-cors==5.0.1
python-dotenv==1.1.0
Werkzeug==3.1.3

# Data Processing and Analysis
pandas==2.2.3
numpy==2.2.5
openpyxl==3.1.5
xlrd==2.0.1
python-docx==1.1.2
pypdf==5.2.0

# Data Visualization
matplotlib==3.10.1
seaborn==0.13.2
plotly==6.0.1
networkx==3.4.2
wordcloud==1.9.4
geopandas==1.0.1

# Natural Language Processing
spacy==3.8.5
transformers==4.46.3
sentence-transformers==4.1.0
scikit-learn==1.5.2

# Database and Storage
pymongo==4.12.0
psycopg2-binary==2.9.10
SQLAlchemy==2.0.40

# Utilities
python-multipart==0.0.20
requests==2.32.3
tqdm==4.67.0
Pillow==11.2.1
opencv-python==4.10.0.84
opencv-contrib-python==4.10.0.84

# Development and Testing
pytest==8.0.2
black==24.2.0
flake8==7.0.0
